Smart Contract
The IDE is, again and again, showing "type-error" for the code despite debugging a lot of times.  We cross-checked it with the code provided at the end of the Module2(CryptoCodeSchool) and the same error was popping up
After coming up to the conclusion, that it is an IDE's error, we are submitting the project.